Set `cmd` and `executable` to `None`
If `executable` is set to any value then SublimeLinter will look for the executable in the system path. If it doesn’t find the executable then the linter will be deactivated.

Now the `livecode-server-path` variable can be set to any path and renaming the `livecode-community-server` executable won’t cause a problem.